+++ src/utils/index.ts
@@ -22,3 +22,16 @@ export const getUserData = (): UserLocalData => {
     sparkUserId: user.sparkUserId,
   };
 };
+
+export const getDataFromLocalStorage = (key : string) => {
+  const data = localStorage.getItem(key); // localStorage에서 데이터를 가져옴
+  if (data) {
+    try {
+      return JSON.parse(data); // JSON으로 파싱하여 반환
+    } catch (error) {
+      console.error("JSON 파싱 오류:", error);
+      return null; // 파싱 실패 시 null 반환
+    }
+  }
+  return null; // 데이터가 없으면 null 반환
+};
